Memory：更改实现方式
