+++ vk.kernel/core/inc/Memory.h
@@ -7,68 +7,18 @@
 #ifndef __MEMORY_H__
 #define __MEMORY_H__
 
-#include "Defines.h"
+#include "Module.h"
 
-///Memory
-class Memory
+/// @brief Memory
+class Memory : public Module
 {
-private:
-	//Structures
-	struct Map 
-	{
-		uint32_t addr;
-		uint32_t size;
-
-		Map(uint32_t addr = 0, uint32_t size = 0):
-			addr(addr),
-			size(size)
-		{}
-	};
-
-	struct MapNode
-	{
-		Map map;
-		MapNode* prev;
-		MapNode* next;
-
-		MapNode(Map map = 0):
-			map(map),
-			prev(NULL),
-			next(NULL)
-		{}
-	};
-
-	//Static constants
-	const static uint8_t  align = 4;
-	const static uint32_t size_of_node = sizeof(MapNode);
-	const static uint32_t kernel_rsvd_heap  = 0x4000;  //10k
-	const static uint32_t kernel_rsvd_stack = 0x3200;  //8k
-
-	//Sram parameters
-	bool isMemReady;
-	uint32_t sram_start;
-	uint32_t sram_ended;
-	uint32_t sram_used;
-	uint32_t sbrk_heap;
-
-	//Members
-	MapNode* head;
-	MapNode* tail;
-	MapNode* curr;
 public:
 	//Methods
-	Memory();
-	~Memory();
-	void Initialize();
-	uint32_t HeapAlloc(uint32_t size);
-	uint32_t StackAlloc(uint32_t size);
-	void Free(uint32_t memory, uint32_t size = 0);
-
-	//Get the max size of sram
-	uint32_t GetSize()  { return sram_ended; }
-
-	//Get the used size of sram
-	uint32_t GetUsed() { return sram_used; }
+	virtual uint32_t HeapAlloc(uint32_t size) = 0;
+	virtual uint32_t StackAlloc(uint32_t size) = 0;
+	virtual void Free(uint32_t memory, uint32_t size = 0) = 0;
+	virtual uint32_t GetSize() = 0;
+	virtual uint32_t GetUsed() = 0;
 };
 
 #endif //!__MEMORY_H__
